+++ R/read_method.R
@@ -5,7 +5,7 @@ NULL
 #' 
 #' Generate text (i.e. html) and read it in the viewer pane of RStudio. If called on
 #' a \code{"partitionBundle"}-object, skip through the partitions contained in the
-#' bundle.
+#' bundle. 
 #' 
 #' To prepare the html output, the method \code{read} will call \code{html} and
 #' \code{as.markdown} subsequently, the latter method being the actual worker. Consult
@@ -13,7 +13,13 @@ NULL
 #' 
 #' The param \code{highlight} can be used to highlight terms. It is expected to be a
 #' named list of character vectors, the names providing the colors, and the vectors
-#' the terms to be highlighted.
+#' the terms to be highlighted. To add tooltips, use the param \code{tooltips}.
+#' 
+#' The method \code{read} is a high-level function that calls the aforementioned methods.
+#' Results obtained through \code{read} can also be obtained through combining these
+#' methods in a pipe using the package \code{magrittr}. That may offer more flexibility,
+#' e.g. to highlight matches for CQP queries. See examples and the documentation for the
+#' different methods to learn more.
 #' 
 #' @param .Object an object to be read (\code{"partition" or "partitionBundle"})
 #' @param meta a character vector supplying s-attributes for the metainformation
@@ -31,9 +37,6 @@ NULL
 #' @param i if \code{.Object} is an object of the classes \code{kwic} or \code{hits},
 #' the ith kwic line or hit to derive a partition to be inspected from
 #' @param type the partition type, see documentation for \code{partition}-method
-#' @param cqp a list of character vectors with regular expressions to
-#'   highlight relevant terms or expressions; the names of the list provide the
-#'   colors
 #' @param cutoff maximum number of tokens to display
 #' @param ... further parameters passed into read
 #' @exportMethod read
@@ -54,14 +57,6 @@ NULL
 #'    highlight = list(yellow = c("Deutschland", "Bundesrepublik"), lightgreen = "Regierung"),
 #'    meta = c("text_name", "text_date")
 #' )
-#' 
-#' all <- partition("PLPRBTTXT", list(text_id = ".*"), regex = TRUE, type = "plpr")
-#' speeches <- as.speeches(
-#'   all, sAttributeDates = "text_date", sAttributeNames = "text_name", gap = 500
-#' )
-#' read(speeches, meta = c("text_date", "text_name"))
-#' migVocab <- count(speeches, query=c("Migration", "Integration", "Zuwanderung"))
-#' read(migVocab, col = "Integration", partitionBundle = speeches)
 #' }
 #' @seealso For concordances / a keword-in-context display, see \code{\link{kwic}}.
 setGeneric("read", function(.Object, ...) standardGeneric("read"))
@@ -71,7 +66,7 @@ setMethod(
   "read", "partition",
   function(
     .Object, meta = NULL,
-    highlight = list(), cqp = FALSE, tooltips = NULL,
+    highlight = list(), tooltips = list(),
     verbose = TRUE, cpos = TRUE, cutoff = getOption("polmineR.cutoff"), 
     template = getTemplate(.Object),
     ...
@@ -81,62 +76,34 @@ setMethod(
       meta <- if (is.null(templateMeta)) names(.Object@sAttributes) else templateMeta
     }
     stopifnot(all(meta %in% sAttributes(.Object@corpus)))
-    if (any(cqp)) cpos <- TRUE
-    fulltextHtml <- html(.Object, meta = meta,  cpos = cpos, cutoff = cutoff,  template = template, ...)
+    doc <- html(.Object, meta = meta,  cpos = cpos, cutoff = cutoff,  template = template, ...)
     
-    # if (length(highlight) > 0) {
-    #   if (length(cqp) == 1){
-    #     if (cqp == FALSE){
-    #       .message("highlighting regular expressions", verbose = verbose)
-    #       htmlDoc <- highlight(htmlDoc, highlight = highlight, tooltips = tooltips)
-    #     } else if (cqp){
-    #       .message("highlighting CQP queries", verbose = verbose)
-    #       htmlDoc <- highlight(object, htmlDoc, highlight = highlight, tooltips = tooltips)
-    #     }
-    #   } else if (length(cqp) == length(highlight)){
-    #     if (any(!cqp)){
-    #       htmlDoc <- highlight(
-    #         htmlDoc,
-    #         highlight = highlight[which(cqp == FALSE)],
-    #         tooltips = tooltips[which(cqp == FALSE)]
-    #       )
-    #     }
-    #     if (any(cqp)){
-    #       htmlDoc <- highlight(
-    #         object, htmlDoc,
-    #         highlight = highlight[which(cqp == TRUE)],
-    #         tooltips = tooltips[which(cqp == TRUE)]
-    #       )
-    #     }
-    #   } else {
-    #     message("length of cqp needs to be 1 or identical with the length of highlight")
-    #   }
-    # }
-    
-    if (require("htmltools", quietly = TRUE)){
-      if (interactive()) htmltools::html_print(fulltextHtml)  
-    } else {
-      warning("package htmltools required, but not available")
+    if (length(highlight) > 0) {
+      doc <- highlight(doc, highlight = highlight)
+      if (length(tooltips) > 0){
+        doc <- tooltips(doc, tooltips = tooltips)
+      }
     }
+    doc
   })
 
 #' @rdname read-method
-setMethod("read", "partitionBundle", function(.Object, highlight = list(), cqp = FALSE, cpos = FALSE, ...){
+setMethod("read", "partitionBundle", function(.Object, highlight = list(), cpos = TRUE, ...){
   for (i in 1:length(.Object@objects)){
-    read(.Object@objects[[i]], highlight = highlight, cqp = cqp, cpos = cpos, ...)
+    read(.Object@objects[[i]], highlight = highlight, cpos = cpos, ...)
     key <- readline("Enter 'q' to quit, any other key to continue. ")
     if (key == "q") break
   }
 })
 
 #' @rdname read-method
-setMethod("read", "data.table", function(.Object, col, partitionBundle, cqp = FALSE, highlight = list(), cpos = FALSE, ...){
+setMethod("read", "data.table", function(.Object, col, partitionBundle, highlight = list(), cpos = FALSE, ...){
   stopifnot(col %in% colnames(.Object))
   DT <- .Object[which(.Object[[col]] > 0)]
   partitionsToGet <- DT[["partition"]]
   if (col == "TOTAL") col <- colnames(.Object)[2:(ncol(.Object)-1)]
   toRead <- as.bundle(lapply(partitionsToGet, function(x) partitionBundle@objects[[x]]))
-  read(toRead, highlight = list(yellow = col), cqp=cqp, ...)
+  read(toRead, highlight = list(yellow = col), ...)
 })
 
 #' @rdname read-method